Delete User password confirmation (#91)
* Password confirmation on user deletetion inertia

* Livewire support

* Typo fix

* code style

* update comment

+++ src/Http/Controllers/Inertia/CurrentUserController.php
@@ -5,6 +5,8 @@ namespace Laravel\Jetstream\Http\Controllers\Inertia;
 use Illuminate\Contracts\Auth\StatefulGuard;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Validation\ValidationException;
 use Laravel\Jetstream\Contracts\DeletesUsers;
 
 class CurrentUserController extends Controller
@@ -18,6 +20,12 @@ class CurrentUserController extends Controller
      */
     public function destroy(Request $request, StatefulGuard $auth)
     {
+        if (! Hash::check($request->password, $request->user()->password)) {
+            throw ValidationException::withMessages([
+                'password' => [__('This password does not match our records.')],
+            ])->errorBag('deleteUser');
+        }
+
         app(DeletesUsers::class)->delete($request->user()->fresh());
 
         $auth->logout();
