+++ src/Http/Livewire/DeleteUserForm.php
@@ -4,6 +4,8 @@ namespace Laravel\Jetstream\Http\Livewire;
 
 use Illuminate\Contracts\Auth\StatefulGuard;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Validation\ValidationException;
 use Laravel\Jetstream\Contracts\DeletesUsers;
 use Livewire\Component;
 
@@ -17,6 +19,27 @@ class DeleteUserForm extends Component
     public $confirmingUserDeletion = false;
 
     /**
+     * The user's current password.
+     *
+     * @var string
+     */
+    public $password = '';
+
+    /**
+     * Confirm that the user would like to delete their account.
+     *
+     * @return void
+     */
+    public function confirmDelete()
+    {
+        $this->password = '';
+
+        $this->dispatchBrowserEvent('confirming-delete-user');
+
+        $this->confirmingUserDeletion = true;
+    }
+
+    /**
      * Delete the current user.
      *
      * @param  \Laravel\Jetstream\Contracts\DeletesUsers  $deleter
@@ -25,6 +48,14 @@ class DeleteUserForm extends Component
      */
     public function deleteUser(DeletesUsers $deleter, StatefulGuard $auth)
     {
+        $this->resetErrorBag();
+
+        if (! Hash::check($this->password, Auth::user()->password)) {
+            throw ValidationException::withMessages([
+                'password' => [__('This password does not match our records.')],
+            ]);
+        }
+
         $deleter->delete(Auth::user()->fresh());
 
         $auth->logout();
